+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @abloatai/cli
 
+## 0.44.0
+
+### Patch Changes
+
+- @abloatai/transaction@0.44.0
+- @abloatai/humans@0.44.0
+
 ## 0.43.0
 
 ### Minor Changes