+++ packages/humans/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @abloatai/humans
 
+## 0.44.0
+
+### Patch Changes
+
+- @abloatai/transaction@0.44.0
+
 ## 0.43.0
 
 ### Patch Changes